[query] `HailContext` is vanquished
