+++ hail/hail/src/is/hail/utils/package.scala
@@ -18,7 +18,7 @@ import java.net.{URI, URLClassLoader}
 import java.security.SecureRandom
 import java.text.SimpleDateFormat
 import java.util
-import java.util.{Base64, Date}
+import java.util.{Base64, Date, Properties}
 import java.util.concurrent._
 import java.util.concurrent.atomic.AtomicBoolean
 
@@ -28,7 +28,7 @@ import org.apache.commons.lang3.StringUtils
 import org.apache.hadoop.fs.PathIOException
 import org.apache.hadoop.mapred.FileSplit
 import org.apache.hadoop.mapreduce.lib.input.{FileSplit => NewFileSplit}
-import org.apache.log4j.Level
+import org.apache.log4j.{Level, LogManager, PropertyConfigurator}
 import org.apache.spark.{Partition, TaskContext}
 import org.apache.spark.sql.Row
 import org.json4s.{Extraction, Formats, JObject, NoTypeHints, Serializer}
@@ -113,6 +113,42 @@ package object utils
   type UtilsType = this.type
 
   def utilsPackageClass = getClass
+
+  val LogFormat: String =
+    "%d{yyyy-MM-dd HH:mm:ss.SSS} %c{1}: %p: %m%n"
+
+  def configureLogging(logFile: String, quiet: Boolean, append: Boolean): Unit = {
+    org.apache.log4j.helpers.LogLog.setInternalDebugging(true)
+    org.apache.log4j.helpers.LogLog.setQuietMode(false)
+    val logProps = new Properties()
+
+    // uncomment to see log4j LogLog output:
+    // logProps.put("log4j.debug", "true")
+    logProps.put("log4j.rootLogger", "INFO, logfile")
+    logProps.put("log4j.appender.logfile", "org.apache.log4j.FileAppender")
+    logProps.put("log4j.appender.logfile.append", append.toString)
+    logProps.put("log4j.appender.logfile.file", logFile)
+    logProps.put("log4j.appender.logfile.threshold", "INFO")
+    logProps.put("log4j.appender.logfile.layout", "org.apache.log4j.PatternLayout")
+    logProps.put("log4j.appender.logfile.layout.ConversionPattern", LogFormat)
+
+    if (!quiet) {
+      logProps.put("log4j.logger.Hail", "INFO, HailConsoleAppender, HailSocketAppender")
+      logProps.put("log4j.appender.HailConsoleAppender", "org.apache.log4j.ConsoleAppender")
+      logProps.put("log4j.appender.HailConsoleAppender.target", "System.err")
+      logProps.put("log4j.appender.HailConsoleAppender.layout", "org.apache.log4j.PatternLayout")
+    } else
+      logProps.put("log4j.logger.Hail", "INFO, HailSocketAppender")
+
+    logProps.put("log4j.appender.HailSocketAppender", "is.hail.utils.StringSocketAppender")
+    logProps.put("log4j.appender.HailSocketAppender.layout", "org.apache.log4j.PatternLayout")
+
+    LogManager.resetConfiguration()
+    PropertyConfigurator.configure(logProps)
+  }
+
+  def logHailVersion(): Unit =
+    info(s"Running Hail version $HAIL_PRETTY_VERSION")
 
   def getStderrAndLogOutputStream[T](implicit tct: ClassTag[T]): OutputStream =
     new TeeOutputStream(new LoggerOutputStream(log, Level.ERROR), System.err)
